init commit

+++ src/main/java/com/example/demo/service/service/impl/NoteServiceImpl.java
@@ -1,8 +1,11 @@
 package com.example.demo.service.service.impl;
 
 import com.example.demo.data.entity.NoteEntity;
+import com.example.demo.data.entity.UserEntity;
+import com.example.demo.data.projection.NoteWithUserNameProj;
 import com.example.demo.data.repository.NoteRepository;
 import com.example.demo.service.dto.NoteDto;
+import com.example.demo.service.dto.NoteWithUsernameDto;
 import com.example.demo.service.exception.NoteNotFoundException;
 import com.example.demo.service.mapper.NoteMapper;
 import com.example.demo.service.service.NoteService;
@@ -10,6 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
@@ -25,25 +29,45 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public List<NoteDto> listAll() {
-        return noteMapper.toNoteDtos(noteRepository.getAll());
+        return noteMapper.toNoteDtos(noteRepository.findAll());
+    }
+
+    @Override
+    public List<NoteWithUsernameDto> listAllUserNotes(Long userId) {
+        return noteMapper.toNoteWithUsernameDtoList(noteRepository.findWithUsername(userId));
+    }
+
+    @Override
+    public List<NoteWithUsernameDto> listAllUserNotes1(Long userId) {
+        return noteMapper.toNoteWithUsernameDtoList1(noteRepository.findWithUser(userId));
     }
 
     @Override
     public NoteDto add(NoteDto note) {
         NoteEntity entity = noteMapper.toNoteEntity(note);
         entity.setId(null);
+        entity.setUser(new UserEntity(note.getUserId()));
+        entity.setCreatedDate(LocalDate.now());
+        entity.setLastUpdatedDate(LocalDate.now());
         return noteMapper.toNoteDto(noteRepository.save(entity));
     }
 
     @Override
     public List<NoteDto> addAll(Collection<NoteDto> notes) {
-        Collection<NoteEntity> savedNotes = noteRepository.saveAll(noteMapper.toNoteEntities(notes));
-        return noteMapper.toNoteDtos(savedNotes);
+        Collection<NoteEntity> notesForSave = noteMapper.toNoteEntities(notes);
+        notesForSave.forEach(note -> {
+            note.setCreatedDate(LocalDate.now());
+            note.setLastUpdatedDate(LocalDate.now());
+        });
+        return noteMapper.toNoteDtos(noteRepository.saveAll(notesForSave));
     }
 
     @Override
-    public void deleteById(UUID id) throws NoteNotFoundException {
-        getById(id);
+    public void deleteById(UUID id, Long userId) throws NoteNotFoundException {
+        NoteDto note = getById(id);
+        if (!note.getUserId().equals(userId)) {
+            throw new NoteNotFoundException(id);
+        }
         noteRepository.deleteById(id);
     }
 
@@ -53,12 +77,13 @@ public class NoteServiceImpl implements NoteService {
             throw new NoteNotFoundException();
         }
         getById(note.getId());
+        note.setLastUpdatedDate(LocalDate.now());
         noteRepository.save(noteMapper.toNoteEntity(note));
     }
 
     @Override
     public NoteDto getById(UUID id) throws NoteNotFoundException {
-        Optional<NoteEntity> optionalNote = noteRepository.getById(id);
+        Optional<NoteEntity> optionalNote = noteRepository.findById(id);
         if (optionalNote.isPresent()) {
             return noteMapper.toNoteDto(optionalNote.get());
         } else {
